requirements.txt
oyaml
pyzmq
pytest
